LayoutTests/accessibility/mac/selected-text-range-for-empty-textarea-expected.txt


This tests that the AXSelectedTextRange for a text area will be 0, 0 instead of nil.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.selectedTextRange is '{0, 0}'
PASS successfullyParsed is true

TEST COMPLETE
